Promote header-block validation into dg_xch_core
Promote get_block_challenge, header_block_to_sub_block_record,
get_signage_point_vdf_info, and validate_unfinished/finished_header_block
into dg_xch_core, driven by the weight-proof recent-chain validator.

Block records are looked up through a concrete &HashMap<Bytes32, BlockRecord>
instead of a BlockRecordProvider trait. The VDF and proof-of-space seam stays
a trait (HeaderValidationVerifier) to break the core -> vdf/pos dependency
cycle. Shared missing/rejected error constructors live in consensus, and the
mainnet weight-proof fixture loader is a shared tests/common helper.
