+++ core/src/blockchain/class_group_element.rs
@@ -1,6 +1,9 @@
-use crate::blockchain::sized_bytes::Bytes100;
+use crate::blockchain::sized_bytes::{Bytes32, Bytes100};
+use crate::utils::hash_256;
 use dg_xch_macros::ChiaSerial;
+use dg_xch_serialize::{ChiaProtocolVersion, ChiaSerialize};
 use serde::{Deserialize, Serialize};
+use std::io::Error;
 
 #[derive(ChiaSerial, Copy, Clone, PartialEq, Eq, Serialize, Deserialize, Debug)]
 pub struct ClassgroupElement {
@@ -16,5 +19,18 @@ impl ClassgroupElement {
         Self {
             data: Bytes100::from(bytes),
         }
+    }
+
+    /// chia `std_hash(bytes(output))`. The consensus hash of this class-group element (a VDF output):
+    /// sha256 over its streamable encoding. Used as the signage-point challenge derived from a
+    /// signage-point VDF output during header validation. As a blockchain (not network) type its encoding
+    /// — hence this hash — is independent of the negotiated protocol version.
+    ///
+    /// # Errors
+    /// Returns an error if the streamable encoding of the element fails.
+    pub fn hash(&self) -> Result<Bytes32, Error> {
+        Ok(Bytes32::from(hash_256(
+            self.to_bytes(ChiaProtocolVersion::default())?,
+        )))
     }
 }
