+++ core/src/consensus/mod.rs
@@ -1,13 +1,31 @@
 pub mod block_generator;
+pub mod block_header_validation;
 pub mod block_rewards;
 pub mod coinbase;
 pub mod constants;
 pub mod deficit;
 pub mod difficulty_adjustment;
+pub mod full_block_to_block_record;
 pub mod generator_puzzles;
+pub mod get_block_challenge;
 pub mod make_sub_epoch_summary;
 pub mod overrides;
 pub mod pot_iterations;
+pub mod vdf_info_computation;
+
+use crate::blockchain::sized_bytes::Bytes32;
+use std::io::{Error, ErrorKind};
 
 pub const CREATE_COIN_COST: u64 = 1_800_000;
 pub const AGG_SIG_COST: u64 = 1_200_000;
+
+pub(crate) fn missing(hash: Bytes32) -> Error {
+    Error::new(
+        ErrorKind::NotFound,
+        format!("block record not found: {hash}"),
+    )
+}
+
+pub(crate) fn rejected(msg: &'static str) -> Error {
+    Error::new(ErrorKind::InvalidData, msg)
+}
